+++ lib/src/screens/home.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class HomeScreen extends StatelessWidget {
+  const HomeScreen({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text('ホーム')),
+      body: const Center(
+        child: Text('ホーム'),
+      ),
+    );
+  }
+}
