+++ lib/src/screens/search.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class SearchScreen extends StatelessWidget {
+  const SearchScreen({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text('検索')),
+      body: const Center(
+        child: Text('検索'),
+      ),
+    );
+  }
+}
