+++ lib/src/screens/shop.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class ShopScreen extends StatelessWidget {
+  const ShopScreen({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text('ショップ')),
+      body: const Center(
+        child: Text('ショップ'),
+      ),
+    );
+  }
+}
